Add files via upload
This is the excel shit that contains all the pixel counts of the heritage properties that were assessed from 1985 to 2015. The excel contains a unique ID that relates to the property in the excel file name called Unique_ID_and_property_names. These properties can be joined together using the IDs